+++ install.sh
@@ -81,6 +81,7 @@ Options:
                               (Default build type is Release)
 
   --cmake-arg <argument>      Forward the given argument to CMake when configuring the build.
+  --rm-legacy-include-dir     Remove legacy include dir Packaging added for file/folder reorg backward compatibility.
 EOF
 }
 
@@ -330,6 +331,7 @@ optimal=true
 cleanup=false
 build_sanitizer=false
 build_codecoverage=false
+build_freorg_bkwdcomp=true
 unset architecture
 unset rocblas_dir
 unset rocsolver_dir
@@ -343,7 +345,7 @@ declare -a cmake_client_options
 # check if we have a modern version of getopt that can handle whitespace and long parameters
 getopt -T
 if [[ $? -eq 4 ]]; then
-  GETOPT_PARSE=$(getopt --name "${0}" --longoptions help,install,package,clients,clients-only,dependencies,cleanup,debug,hip-clang,codecoverage,relwithdebinfo,build_dir:,rocblas_dir:,rocsolver_dir:,lib_dir:,install_dir:,architecture:,static,relocatable,no-optimizations,docs,address-sanitizer,cmake-arg: --options hipcdgsrnka: -- "$@")
+  GETOPT_PARSE=$(getopt --name "${0}" --longoptions help,install,package,clients,clients-only,dependencies,cleanup,debug,hip-clang,codecoverage,relwithdebinfo,build_dir:,rocblas_dir:,rocsolver_dir:,lib_dir:,install_dir:,architecture:,static,relocatable,no-optimizations,docs,address-sanitizer,cmake-arg:,rm-legacy-include-dir --options hipcdgsrnka: -- "$@")
 else
   echo "Need a new version of getopt"
   exit 1
@@ -426,6 +428,9 @@ while true; do
         shift ;;
     -k|--relwithdebinfo)
         build_type=RelWithDebInfo
+        shift ;;
+    --rm-legacy-include-dir)
+        build_freorg_bkwdcomp=false
         shift ;;
     --cmake-arg)
         cmake_common_options+=("${2}")
@@ -573,6 +578,12 @@ fi
 
 if [[ "${build_library}" == false ]]; then
   cmake_client_options+=('-DBUILD_LIBRARY=OFF')
+fi
+
+if [[ "${build_freorg_bkwdcomp}" == true ]]; then
+  cmake_common_options+=('-DBUILD_FILE_REORG_BACKWARD_COMPATIBILITY=ON')
+else
+  cmake_common_options+=('-DBUILD_FILE_REORG_BACKWARD_COMPATIBILITY=OFF')
 fi
 
 rocm_rpath=""