+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Full documentation for rocSOLVER is available at [rocsolver.readthedocs.io](http
 ## (Unreleased) rocSOLVER
 ### Added
 - Added tests for multi-level logging
+- File/Folder Reorg
+  - Added File/Folder Reorg Changes with backward compatibility support using ROCM-CMAKE wrapper functions.
 
 ### Optimized
 ### Changed